+++ src/components/Register/Finished/Finished.module.css
@@ -3,4 +3,5 @@
   color: white;
   border-radius: 40px;
   margin-top: 44px;
+  padding: 12px 32px;
 }
