+++ src/components/Settings/ManageWallets/ManageWallets.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const ManageWallets = () => {
+  return <div>ManageWallets</div>
+}
+
+export default ManageWallets
